Menu (#23)
* Add button for filtering out non-event-exclusive items

Got a little carried away; this really should be multiple commits.  This commit implements a first draft of the state structure and dispatching that will be needed for several of the filtering options I'd like to eventually offer to users (only event-exclusive will come first, but I'm planning ahead for the rest).  It also hides the item if it's not being calculated in the bronze total, as well as hiding the vendor if it has no unhidden items.

* Fix conditional logic for news icon animation

* Add menu & move non-event-exclusive toggle into menu

* Add filter for obtained items

* Move item filtering logic into custom hook

* Implement headlessUI switch components

* Componentize menu options

* Improve menu styling

* Hide vendor if all its items are ignored